LSTM, GRU, Transformer, InceptionTime

+++ src/rcpl/dataset.py
@@ -8,7 +8,7 @@ from rcpl.rcpl import Experiment
 class RCLPDataset(Dataset):
     def __init__(self, exp: Experiment, purpose='train', dataset_dir=None):
         self.x = torch.from_numpy(np.load(dataset_dir / f'{purpose}_x.npy'))
-        self.y = torch.from_numpy(np.load(dataset_dir / f'{purpose}_y.npy'))
+        self.y = torch.from_numpy(np.load(dataset_dir / f'{purpose}_y.npy')).float()
         self.epsp = exp.epsp
 
     def __getitem__(self, index):
